+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 3994129 - New VERSION.md
+| Author: fjrodafo - Thu, 13 Feb 2025 00:08:58 +0000
+| 
 * Commit: fd53aa8 - Update FUNDING.yml
 | Author: fjrodafo - Wed, 12 Feb 2025 23:01:40 +0000
 |